nut and lanplay.r3n3 are going to be discontinued
Cause of that i would shut them down from the public list that no one gets confused if i sometimes turn it on if i play with friends in my country :)

it was a great time hosting i meet many great people but for now its time for me too walk on.
i try to learn to program for my own maybe i can give back something back to the community.

for that i discontinue hosting publicly to get focusd to learn some new skillsets.

Thanks to all of you Greatwizard, Space and Cubuss it was a great time. im glad this code was open source the hole time.
i was reading into it and got engaged with Developing and now i wanna walk on and learn how to do such awesome stuff like you guys did.

Peace out and hopefully we see us in future.