Update Download.vue
removed individual links/versions and linked to the readme.md version-table from the ldn_mitm github. a recent PR on ldn_mitm now links to each version of ldn_mitm with info shown for which AMS will work with it. This should clean this page up a bit while also showing up to date compatibility information.