Update communities.json
our leadership closed the old community due to the fact that the administrator left his kingdom and we were unable to contact him. In addition, there was a dead server, and we built a new one with active players as well as a new administration. please change the action of the Discord server. Also, please add our community to the Russian social network vkontakte. Thank you! There we hold contests and tournaments.